refactor(store): fix path
